Changed datatypes for backend, added controller method to get a book by id

+++ MSA-Phase-2/MSA-Phase-2/Repositories/Abstract/IBookRepository.cs
@@ -5,12 +5,11 @@ namespace MSA_Phase_2.Repositories
     public interface IBookRepository
     {
         Task<IEnumerable<Book>> GetAllBooksAsync();
-        Task<Book> GetBookByIdAsync(long id);
+        Task<Book> GetBookByIdAsync(int id);
         Task AddBookAsync(Book Bookbook);
         Task UpdateBookAsync(Book book);
-        Task DeleteBookAsync(long id);
-        Task<bool> BookExistsAsync(long id);
+        Task DeleteBookAsync(int id);
+        Task<bool> BookExistsAsync(int id);
         Task BulkAddBooksAsync(IEnumerable<Book> books);
-
     }
 }
